TCPMv2: fix force power role when powered cable w/o sink attached (EMCA)
Observed that while testing firmware_PDConnect.

1. servo-v4p1 as SRC and DUT as SNK
2. enforce the DUT to force source
3. the DUT detects Ra(0.1V)/Open(3.1V) servo-v4p1 EMCA mode, and leaves
   LPM
4. the TCPMv2 transit to Auto-Toggle mode
5. DUT starts toggling starting as presenting Rd
6. servo-v4p1 recognize the CC level as Rd, sourcing 5v
7. DUT though presenting Rd, but it's DRP role is SRC, so failed to
   connect

This CL is to fix the step 4 and 5.

The Ra/Open pair is recognized as powered cable without sink attached.
In this case, it's treated the same as Open/Open, so it should stay
at the status as is regards to the power role setting. That is

When we have asked the TCPM to force source, we should not force it into
the auto-toggle mode under any case.

BUG=b:268077715
TEST=- On steelix (it81202 TCPC), voltorb (rt1718s, and anx3447)
     test with servo-v4p1 with the following test:
     1. v4p1 SRC (+-EMCA), DUT SNK -> SRC -> SNK
     2. v4p1 SNK (+-EMCA), DUT SRC -> SNK -> SRC
     - Pass firmware_PDConnect on voltorb
BRANCH=none

Change-Id: Ifba0cfb79652a699a274b0086bb0d9b5f888d7a8
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>