mt8186: fix sysjump and hibernate actions
1. fix sysjump will boot the system with AP IDLE
1. fix boot with AP OFF
2. fix boot from hibernate

The recent change which make the system staying at S5 for 10 seconds
when AP power off. When sysjump requested, the RW part will
automatically boot the system from S5 due to the initial in_exiting_off
flag defaults to true. Since the exiting_off status is controlled by the
power_chipset_init, we don't need the flag defaults to true on boot.

Also, change the initial power state S5 to G3. They are the same in
rail-wise on mt8186, but not in mt8188. Assign it as G3 to match
mt8188 initial GPIO and rail states. When the AC on under hiberante,
or the EC boots with AP OFF flag, this ensures the system stay at G3.

BUG=b:274051287 b:274063396 b:274368558
TEST=pass firmware_ECPowerButton
TEST=pass firmware_ECWakeFromULP
TEST=dut-control cold_reset:on sleep:1 cold_reset:off, boot to S0
TEST=dut-control power_state:rec; dut-control power_key:tab. stay at
     S5->G3
TEST=hibernate, and AC on. The EC stays at G3.
BRANCH=none

Change-Id: I253f4956665a6a0edc2fe59c5a8a90a7c07a6180
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>