zephyr: ppc: support multiple altnernates per port
Add support for multiple PPC alternates per port.  This change compiles
and creates an array called ppc_alternates[], that is initialized with
all PPC nodes found in the tree that are enabled and have the "is-alt"
property set.

Next change will define discrete ppc_config_t structures for each
alternate so the board code can lookup the correct alternate based on
the node id.

Change-Id: Iddfb129d81268f255ba018923218eb96c994f72c
Signed-off-by: Keith Short <keithshort@chromium.org>