rex: Remove CR50 reset workaround config
Adapted from chromium:3933588

This commit removes `CONFIG_PLATFORM_EC_BOARD_RESET_AFTER_POWER_ON` from
rex. This config enables a brief delay as a workaround for a Cr50 issue
that triggers a second reset of the EC shortly after it is reset. Ti50
does not have this issue and this delay prevents GPIOs from being
configured before the jump to RW. RW does not attempt to configure GPIOs
again as it assumes that they were correctly configured by RO. The
misconfigured GPIOs cause the TCPC to be kept in reset which ultimately
causes a boot loop.

BRANCH=none
BUG=b:270657357
TEST=rex can boot to welcome screen on AC plugin (no battery)

Change-Id: I6da12616b7e31193cc1a64c0ac3d801ea77db39a
Signed-off-by: Caveh Jalali <caveh@chromium.org>