chip/stm32: Eliminate support for multiple USB->SPI instance
For some reason, all methods in usb_spi.c pass along pointers to config
and state, as if it made sense to have multiple USB interfaces each
implementing their own instance of the SPI->USB bridge protocol.  This
makes little sense, as a single instance is able to manipulate multiple
SPI busses.

This CL eliminates the pointer parameters, for a tiny performance
improvement, and code simplification.

BUG=273601311
TEST=TBD

Change-Id: Ifbfdb18b4a2ba656732e4eb143de9f678416067a