zmake: add check for driver GPIOs
The named-gpios node is used to configure the initial state of all
GPIOs. However, as drivers are moved upstream the GPIO configuration
will be done by the driver.

Add a configure time check that validates if a GPIO is specified in
named-gpios and as a property to a driver, that the flags match. This
will help prevent regressions in the future.

BUG=none
BRANCH=none
TEST=zmake configure -a. Verify no GPIO flag mismatches found.

Change-Id: Ibb144dcd8734872c3aeb2bd0995c473b9a59490c
Signed-off-by: Keith Short <keithshort@chromium.org>