cr50: add a factory config space
Add a 64 bit write once factory config space to info1. If the factory
has something they want to configure, they can use part of the space to
store that configration.

Right now nothing in cr50 uses the factory config space. If we need to
modify cr50 behavior based on the space value, we can add functionality
later. The factory just needs to set the bit in the factory config.

BUG=b:214065944
TEST=manual
	Clear the Board ID
	set the config to 0. Verify it does nothing
	gsctool -a --factory_config 0

	Set the flags
	gsctool -ai 0xffffffff:0x10
	Set the config to something
	gsctool -a --factory_config 0x12345678cafecafe
	# Set it to the same thing. Verify cr50 returns EC_SUCCESS.
	gsctool -a --factory_config 0x12345678cafecafe
	[40.114944 write_factory_config: ok.]
	Try to set it to something else. Verify it's rejected because
	the space is set.
	gsctool -a --factory_config 0xcafe
	[43.331302 write_factory_config: factory cfg already programmed]

	Set the Board ID Type
	gsctool -ai ZZCR:0x10
	Try to set the config again. Verify it's rejected because the
	board id type is set.
	gsctool -a --factory_config 0x12345678cafecafe
	Factory config failed. (7)

Change-Id: Ie816ebffcf6c24ad94bbcd2dc2f0c3936caafb11
Signed-off-by: Mary Ruthven <mruthven@chromium.org>