chip/stm32: Eliminate run-time flags to usb_spi.c
The USB_SPI_CONFIG() macro allowed a set of "run-time" flags to be given
by board.c.  However, these flags have to be known at compile time, and
always are.  They can more conveniently be given through a CONFIG_
define.

With the above, and a little tweak in hammer/board.h (mimicking what was
already done for I2C), then every invocation of the USB_SPI_CONFIG()
across the codebase is identical, setting the stage for moving the
declarations into usb_spi.c, and eliminating the macro altogether.

BUG=b:273601311
TEST=none

Change-Id: Id565338238a7c21f7e2c6c24b7c4e3d7648482f6