rollback: Initialize first rollback region if no valid region was found
In CrosEC, initial values for first rollback region are provided in
the firmware image (see common/firmware_image.lds.S file). In Zephyr,
it's not possible to implement it the same way because the firmware
image is assembled in a different way.

This problem was solved by checking if there is a valid rollback region
on boot.

BUG=b:277029648
BRANCH=none
TEST=Compile Zephyr FPMCU firmware for bloonchipper. Check if
     'rollbackinfo' reports the following:

     fpmcu:~$ rollbackinfo
     rollback minimum version: 0
     RW rollback version: 0
     rollback 0: 00000000 00000000 0b112233 [00..00] *
     rollback 1: ffffffff ffffffff ffffffff [ff..ff]

Change-Id: I9812e7d123061fdeec7b283f37d2d12fab4f8424
Signed-off-by: Patryk Duda <pdk@semihalf.com>