zephyr: add checks for driver GPIOs
The named-gpios node is used to configure the initial state of all
GPIOs. However, as drivers are moved upstream the GPIO configuration
will be done by the driver.

Add a configure time check that validates if a GPIO is specified in
named-gpios and as a property to a driver, that the flags match. This
will help prevent regressions in the future.

BUG=b:271135255
BRANCH=none
TEST=zmake configure -a
TEST=./twister -b

Change-Id: I45aa14bf93e74877082a2a674b391cfab56379e4
Signed-off-by: Keith Short <keithshort@chromium.org>