nissa: test project-level sub_board.c
This defines a Nissa-like board for unit testing and adds tests to exercise
the project's sub_board.c, in particular that GPIO configuration is
correct.

A small change is made to make it possible for the tests to reset some
global values that are set by board code on initialization; there is no
functional change to board code.

BUG=b:271118112
TEST=./twister -ciC -T zephyr/test/nissa
BRANCH=nissa

Change-Id: I8edfddbf663761f5c72c4d680cc056ee7d46d2f8
Signed-off-by: Peter Marheine <pmarheine@chromium.org>