PRESUBMIT: Remove checkpatch from presubmit
Now that clang-format is enabled and enforced by default, we should
stop running checkpatch on EC code during presubmit.

checkpatch often conflicts with clang-format, and it's much easier to
have a single source-of-truth for formatting, so that IDEs and other
tools can format automatically.

The EC headers that are included in the Linux kernel are still checked
with checkpatch through util/linux_ec_commands_h_check.sh.

BRANCH=none
BUG=b:276785070
TEST=none

Change-Id: I87e4a6527ed2ab1c2bf158c871dd5b71ae4a8e9f
Signed-off-by: Tom Hughes <tomhughes@chromium.org>