mt8186,mt8188: check the holder of AP_RST_ODL
To distinguish the AP shutdown, and the AP reset held by GSC/Servo,
we use the SYS_RST_ODL as the reference.

If AP_EC_SYSRST_ODL is asserted, but SYS_RST_ODL is not, this is a
normal shutdown.
If AP_EC_SYSRST_ODL is asserted, and so is SYS_RST_ODL, the AP reset
is held by GSC or Servo. In this case, we should let the power state
stay at S0 to prevent from an unexpected shutdown.

BUG=b:276229973
TEST=pass firmware_CorruptMinios
TEST=dut-control warm_reset:on sleep:1 warm:reset:off, it stays at S0
BRANCH=none

Change-Id: I9b69caa0d15e6e58e7f11ad1079536bc1954b5ce
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>