pinmap: migrate code to Arbitrage
This commit removes pinmap code from the chromium codebase because it
has been migrated to go/arbitrage.

In order to use this code going forward, users should follow the steps
in go/arbitrage to setup Arbitrage and the steps in
https://team.git.corp.google.com/cros-arbitrage/arbitrage#zephyr-pinmap-generation
to generate pinmaps.

In order to develop this code going forward, developers should follow
the steps in
https://team.git.corp.google.com/cros-arbitrage/arbitrage/+/HEAD/documentation/dev_guide.md.

For any questions, contact the Arbitrage team at
arbitrage-eng@google.com.

Bug: 274938064
Change-Id: I2948ee2245236d95a133758e695e51ecbb4969c9