[WIP] Reduce fan duty oscilations
This is a debugging/testing patch, don't merge it!

Descrase fan dut_step variable to make the fan spin up/spin down
slower.
This should improve behaviour with fans that have a lot of inertia.
Also apply higher body detection in order to get it working with
LIS2DW12 sensor.

Change-Id: I8bf5506f346dcb6fcb78693623b29121159a52fa
Signed-off-by: Kornel Dulęba <korneld@google.com>