zephyr: Add support for board override of fan duty adjustment logic
The generic logic has proven to not work well on whiterun.
In order to mitigate that, while not affecting everyone else,
a custom fan duty adjustment logic needs to be implemented.
The behavior is controlled with CONFIG_CUSTOM_FAN_DUTY_CONTROL,
which will be enabled on whiterun in a another patch.
Some structs has been moved to the fan header, so that they can be
accessed from the custom logic.

BUG=b:276577324
BRANCH=none
TEST=Build test Brya and Skyrim targets

Change-Id: Iff31b6526be83b8b0165c4490bafe26f131e6c84
Signed-off-by: Kornel Dulęba <korneld@google.com>