WIP: winterhold: Implement custom fan duty adjustment logic
It's strongly based on the existing logic from zephyr/shim/src/fan.c.
The main difference is that the fan duty is changed less aggressively,
in order to avoid fan oscillation.
This is a quick fix, at some point it should be replaced by a PID type
of algorithm.

BUG=b:276577324
BRANCH=none
TEST=Verify that the fan doesn't ramp up and down all the time under
     various scenarios, e.g. watching multiple videos on youtube.

Change-Id: I893a7802df8f5271fe1a253272fdcc3149bdf178
Signed-off-by: Kornel Dulęba <korneld@google.com>