whiterun: Adjust CPU fan thermal parameters
The difference between temp_fan_off and temp_fan_max was set to only 6C.
Because of that a one degree difference in CPU temp would cause a
significant RPM difference. This results in noticeable noise variations
when the CPU is hit with just the right load.
Decrease the temp_fan_off by 4C to alleviate this.

BUG=b:276577324
BRANCH=none
TEST=Run webGL aquarium with various number of fish

Change-Id: I72ec924e61a68e61a9c7f9721b7df15d18aaa2ee
Signed-off-by: Kornel Dulęba <korneld@google.com>