whiterun: Adjust CPU fan thermal parameters
The difference between temp_fan_off and temp_fan_max was set to 6°C.
Consequently, a one-degree difference in CPU temperature would cause
a significant change in fan RPM. This results in noticeable noise
variations when the CPU is subjected to a specific load.
To alleviate this issue, decrease the temp_fan_off by 4°C, which,
in turn, reduces the steepness of the temperature-to-RPM relationship.

BUG=b:276577324
BRANCH=none
TEST=Run webGL aquarium with various number of fish

Change-Id: I72ec924e61a68e61a9c7f9721b7df15d18aaa2ee
Signed-off-by: Kornel Dulęba <korneld@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4427797
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>
(cherry picked from commit 0852c76440bade57d203bc0f985de56a548c530c)