whiterun: Adjust CPU fan thermal parameters
The difference between temp_fan_off and temp_fan_max was set to a mere 6°C. Consequently, a one-degree difference in CPU temperature would cause a significant change in fan RPM. This results in noticeable noise variations when the CPU is subjected to a specific load. To alleviate this issue, decrease the temp_fan_off by 4°C, which, in turn, reduces the steepness of the temperature-to-RPM relationship.

BUG=b:276577324
BRANCH=none
TEST=Run webGL aquarium with various number of fish

Change-Id: I72ec924e61a68e61a9c7f9721b7df15d18aaa2ee
Signed-off-by: Kornel Dulęba <korneld@google.com>