Hades: Unblock USB_C1_PPC_SNK_EN on start-up
When a system boots on only AC power (i.e. without a battery or with a
depleted battery), the PPC needs to be set to sink. However
USB_C1_PPC_SNK_EN of ANX7406 is deasserted by default. Hades has a
MOSFET to block this signal so that ANX7406 won't set the PPC to source
on start-up.

This CL sets the GPIO0 of ANX7406 to high on start-up. This unblocks
USB_C1_PPC_SNK_EN so that ANX7406 can take over the PPC control.

BUG=b:269790564
BRANCH=None
TEST=make BOARD=hades

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I644e207f0923c1f5a3111d9a1382defd0e34715f