zmake: compare-builds: Fix project directory
The configure step run at each checkout was still using the project
directory from the user's main source tree.

After the code checkout is complete, update the projects_dir so that the
configure step uses the correct checkout.

BUG=none
BRANCH=none
TEST=zmake -l DEBUG zcompare-builds -a.  Verify that the -S option
passed to cmake is the checkout directory and not the user's active
source directory

Change-Id: Ib5ece61d73a28eecde2b199c684c8c40318c53ee
Signed-off-by: Keith Short <keithshort@chromium.org>