kingler: only reset TCPC with power-on reset flag
Reset the TCPC after a power-on restart, and clear any explicit
PD contract so that the previous state is not restored. The TCPC
reset is not needed on starting up with EFS2 enabled since the
contract will be established after entering RW.

This fixes that while the EC soft-reset is issued, the TCPC
will be reset on the starting up, and this will cause a brown-out
if no battery connected.

BUG=b:276661970
TEST=steelix with AC only, no battery, issue reboot and system boots.
BRANCH=none

Change-Id: I17f9f3122a51a179996e1a950a18887cb2a5c8be
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>