zephyr: make console buffer logging depend on non-minimal logging
The console buffer log backend depends on CONFIG_LOG_OUTPUT, which only
does something if logging is enabled and not in minimal mode. Without
this dependency, enabling this backend while in minimal mode causes a
link-time error because symbols enabled by CONFIG_LOG_OUTPUT are absent.

When in minimal mode log backends are ignored completely anyway, so this
change also helps prevent accidental misconfiguration because it doesn't
make sense to enable a log backend that will never be used.

BUG=b:265781350
TEST=zmake build -a
BRANCH=none

Change-Id: Ib0722102963b44484f07a779b2bbbde72b29b743
Signed-off-by: Peter Marheine <pmarheine@chromium.org>