npcx: watchdog: support the longer watchdog timeout
In the current pre-scalar setting (fixed to 0), the max watchdog timeout
that the user can set is ~8 seconds. It is not enough for some
application requirements such as FPMCU. This commit extends the max
timeout by adjusting the pre-scalar dynamically.

BRANCH=none
BUG=b:273577369
TEST=pass "make buildall -j"
TEST=Change the CONFIG_WATCHDOG_PERIOD_MS > 8 seconds; type console
command "watims ${timeout}"; make sure watchdog panic occurs as
expected.
TEST=Set the CONFIG_WATCHDOG_PERIOD_MS to the default (1600); make sure
WDCNT and WDCP are the same before/after applying this commit.

Change-Id: I11b764811f65f0956593ab6725a619a1a44464c4
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>