zephyr: LED: Refactor LED driver
This CL formats the led-policy devicetree to be more intuitive as part
of the LED driver overhaul. New LED driver now allows multiple LEDs to
be manipulated concurrently instead of consecutively. Also allows for
LED brightness to be adjusted on a separate tick timing from policy
application in preparation for LED transitions CL.

BUG=b:228868616
TEST=twister, battery LED operation verified on Skyrim, Xivu, ectool
     tested on Skyrim, Xivu.
LOW_COVERAGE_REASON=This refactor is in preparation for adding LED
                    transitions. twister tests will be added with the
		    LED transitions CL.

Change-Id: Ia2307682a12690d57062a6b59f831b4fdbfc7279