board/hyperdebug: Use dedicated memory buffer for gpio monitoring
HyperDebug can be instructed to capture edge events on a subset of GPIO
pins.  The current use of `common/shared_mem.h` conflicts with its use
by e.g. SPI.

This CL introduces a dedicated memory area in `hyperdebug/gpio.c`.  For
now, only a single allocation of that area is implemented.  We may want
to allow the called to request a certain buffer size, and split the
available are between buffer for disjoint subsets of GPIO pins being
monitored.

BUG=b:279752946
TEST=opentitantool spi tpm read register,  while gpio monitoring

Change-Id: I9cc9b016b964b4db1cb31a048126deeacfef9b99