board/hyperdebug: Avoid losing GPIO events
`ccprint()` will never block, and will drop data if buffers are full.
This is undesirable when opentitantool is fetching a list of GPIO edge
events.

This CL uses the same approach as for instance the `gpioget` command in
`common/gpio_commands.c`, which may also want to print a long list,
which would not fit in the buffer.

BUG=none
TEST=observe a list with more than 100 events in EC console

Change-Id: Ide059d615c0be32efd0f4aea29815518f4216a16