Rex: Allow VCONN swap when PWR_ALL_SYS_PWRGD is asserted
The current USB PD policy will only allow Vconn swaps when the AP is on.
This causes the EC to send a hard reset to USB PD partners instead of
trying to Vconn swap when the device is booting, leading to device
discovery issues in the kernel. This CL updates the Vconn swap policy to
allow Vconn swaps when PWR_ALL_SYS_PWRGD is asserted. During boot, this
policy prevents a USB PD hard reset from disrupting device discovery in
the cros-ec-typec driver.

BUG=b:274576896
BRANCH=none
TEST=emerge-rex chromeos-base/chromeos-zephyr, ./twister -T
zephyr/test/rex -c, checked that the EC did not hard reset with
connected partners during boot.

Change-Id: Ic9560509f4d7f6161bb001a80ba588f3d5108b13
Signed-off-by: Jameson Thies <jthies@google.com>