WIP: POC: winterhold: PID algorithm PoC
This is only a POC, it's not meant to be merged, or used in production.
The idea here is to implement a generic solution based on this.

Leverage the board_override_fan_control_duty function to implement a PID
PoC on winterhold.
Currently only PI part is used, i.e. the derivative coefficient is set
to 0.
I've found that this configuration yields acceptable results - no fan
oscillation related to fan duty changes was observed.
If the fan RPM changes with this config it's likely due to nature of the
"linear" thermal function approach.

In order to adjust the parameters quickly a new EC command was added -
"setpid". It takes integers as input and then shifts it by 4 decimal
places to get the coefficients.

Change-Id: I6f5f4053551539602f9579a1d54872ab16dcfe9f