cr50: set EPS seed length in NV state upon creation
Currently EPS seed when created during endorsement is immediately stored
to flash and copied to global state of TPM. However, its length in state
is not set. It seems that it was ok as state eventually was reloaded
from NV and size updated. But if TPM receives command shortly after
boot this may not be the case.

BUG=b:278028110,b:262324344
TEST=make; TCG tests

Change-Id: Ifac259ea3817d6d139d9b5ba2af4fc66d07df4ef
Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>