zephyr: Add a new project for native development
The only native-posix project in the tree is described as 'minimal' so
is not suitable for development, where we need a good set of features
enabled.

Add a new 'dev-posix' project for this. It will form the basis for the
charger-task development. It is based on 'minimal-posix'.

BUG=b:218332694
BRANCH=none
TEST=zmake build dev-posix
build/zephyr/dev-posix/build-singleimage/zephyr/zephyr.elf
See that it runs OK

Change-Id: I3e1d4fe2df683a5d587702aadfe40fbd2816e8b2
Signed-off-by: Simon Glass <sjg@chromium.org>