driver/tcpm/nct38xx: Remove write to RO register
`TCPC_REG_DEV_CAP_2` is a RO register according to the TCPC spec and the
NCT38xx datasheet and writes to this register have no effect on real
hardware. The vendor has also confirmed that they are not aware of any
workarounds that would explain this write. Remove it.

BUG=b:280351196
TEST=Built and ran on Skyrim, verified that FRS swap feature is still
enabled in TCPC_REG_CONFIG_EXT_1.

Change-Id: I9364f66c3499a7a49b41c9edad6cd6ef970e6a89