corsola: Improve XHCI interrupt naming and remove redundant KConfig
The `int_usba` interrupt configuration in dts is actually configuration
for XHCI init done interrupt, which is not equivalent to settings for
USB A port. Update the naming and the interrupt handler function name,
and remove the KConfig `VARIANT_CORSOLA_USBA` as this name does not
match to the actual effect and whether to have USBA related settings are
controlled by the usba.dtsi is imported or not.

BUG=b:274723043
TEST=`zmake build <all board> --clobber`, and the USB A port works on
     Krabby proto 1 board.

Change-Id: I400990745804e059f22b02845a11d1367ea079f4