panic_output: Do not use ccprintf to induce stack overflow
The console command to induce a stack overflow crash uses ccprintf,
which acquires a mutex lock internally. The exception handler aborts the
the thread (shell_uart) when the stack overflow is detected, but the
mutex is not released. This causes system safe mode to deadlock.
   
Switch to panic_printf to avoid this deadlock. A more robust long
term fix may be to change Zephyr internals to release mutexes on aborted
threads.

LOW_COVERAGE_REASON=Swapping single function call
BUG=b:278792557
TEST=Able to start system safe mode after a stack overflow on skyrim

Change-Id: Ib7b50487d27e77fdd5b6cec5176060074868dca3