corosla: Improve USB-C port configuration
The number of USB-C port configuration is hard coded in the config file
and is set to 2. This becomes a blocker for boards with different number
of USB-C ports, as the build assertion assumes the number of UBC-C ports
and the number of USB PD ports is euqal, yet USB PD ports is determined
by the dts configuration.
Update the USB-C ports configuration by using macro `LISTIFY` to
generate the UBS-C ports configuration, and also add logic to make
different number of USB-C port compatible with implementation in corsola
projects.

BUG=b:274723043
TEST=`zmake build <all board> --clobber`, and test the dp, usb
     functionality on Krabby proto 1 board.

Change-Id: I04622ec5d6bc4e7f396517e3f8e9b024ced078e8