corsola: Update implementation for no daughter board
CORSOLA_DB_NONE was treated as no daughter board connected, however,
there are 2 cases fall into this category:
1. There is no port 1 design on that board.
2. There is port 1 design on that board, yet it is not connected during
   development.

As we update the implementation of configuring USB-C port in "corosla:
Improve USB-C port configuration", we can defer the case 1 to dts
configuration, so CORSOLA_DB_NONE is now dedicated for case 2. Update
the returned pd port count so that if there is only 1 USB-C port, the
only port still functions normally.

BUG=b:274723043
TEST=By (1) removing port 1 configurations in Krabby,
        (2) setting `VARIANT_CORSOLA_DB_DETECTION` to no in Krabby,
        and build, flash onto Krabby proto 1 board, only port 0 is
        functioning with
           a. Accepting charging.
           b. USB device is working.
           c. DP works with external display (kernel is modified
              accordingly).

Change-Id: I1aadb63a83b452ea8de83309d464918f76c7a511