chgstv2: Always discharge if soc is > sustainer_soc.upper + 1
For the Charge Limit feature, we need to discharge down to the upper
limit for the battery sustainer. The Charge Limit feature requires that
the system never charge above 80%, so if the feature is enabled when the
system is over that, it needs to discharge down to 80%. This compares to
upper + 1 to avoid charge/discharge thrashing.

BUG=b:222619859
TEST=run-sbs_charging_v2

Change-Id: I3f65743144de64411484dd3f2aa6c5fd681a47c9