keyboard: Hold scancode queue when SETLEDS command is expected
When EC sees ALT is released while SEARCH is pressed or SEARCH is
released while ALT is pressed, it'll set a flag (i.e. setleds_coming).
While this flag is set, the EC blocks the scancode queue and waits for
the SETLEDS command.

This flag is cleared when SETLEDS arrives or a timer (i.e.
setleds_coming_deadline) expires. Then, the scancode queue is unblocked.

BUG=b:273636117,b:271948416,b:237981131,b:247795316
TEST=Redrix

Change-Id: I9af596cfae23575d0c44866ad17c57e5cbd9c187