keyboard: Hold scancode queue when SETLEDS command is expected
When EC sees alt and search pressed, it'll set a flag (e.g.
setleds_coming). When one of the two keys is released, EC blocks the
scancode queue and waits for the SETLEDS command.

If EC receives SETLEDS, it returns ACK and unblocks the scancode queue.
If EC doesn't receive SETLEDS before a timer expires, it unblocks the
scancode queue.

BUG=b:273636117
TEST=Redrix

Change-Id: I9af596cfae23575d0c44866ad17c57e5cbd9c187