charger: Make is_full mostly a local variable
This is a property of the charger task and various functions access it.
Within each iteration of the charger loop, it is only changed at most
once.

It seems better to use a local variable for that.

Unfortunately the exported function charge_get_percent() needs access
to this variable. But it seems better to group exported information
together, updating it as needed. Then this information it could
potentially be accessed as a cohesive whole, rather than a collection
of unrelated pieces.

Make a start on this by creating a new local struct to hold what will
be the 'published' information. Update it once around the charger loop.
Update charge_get_percent() to use that version of the information.

Code size on lux, which seems to be the only user of the base code,
increases a bit. This hits it fairly hard since it adds a 5th function
argument to set_base_lid_current(). We could improve that by using
flags instead of three booleans. But it seems best to do that once we
have all the reported state in one place. Note that we get much of the
code space back with the next CL.

This makes no functional change.

BUG=b:218332694
TEST=zmake build dev-posix
Lux adds 136 bytes to RO, less to RW; RAM improves a little
Free space before:
  *** 69624 bytes in flash and 1152 bytes in RAM lux RO ****
  *** 69444 bytes in flash and 1120 bytes in RAM lux RW ****
after:
  *** 69488 bytes in flash and 1152 bytes in RAM lux RO ****
  *** 69476 bytes in flash and 1152 bytes in RAM lux RW ****

Change-Id: Ide34a613e0bab7868d268b6fa1a9db258d3c5945