charger: Make base_connected() a function
The double declaration of this variable is ugly. The only user seems to
be lux, where board_is_base_connected() is implemented simply by
reading from another variable. So there is really advantage to caching
the value in the charger code.

Change base_connected() to a function, where it returns false if there
is no base, otherwise calls the base code to find out.

This makes no functional change.

BUG=b:218332694
TEST=zmake build dev-posix
Check code on lux:
  *** 69552 bytes in flash and 1152 bytes in RAM lux RO ****
  *** 69424 bytes in flash and 1152 bytes in RAM lux RW ****

Change-Id: I14a05a7705fa52c92fa15b0ee233f79ef01d1b53