config: Don't enable LIBCRYPTOC in config.h
In config.h we define CONFIG_LIBCRYPTOC when CONFIG_ROLLBACK_SECRET_SIZE
is defined. This way we are sure that rollback have an access to
always_memset() function, used to securly clear buffers.

However, if BoringSSL library is enabled we use OPENSSL_cleanse()
function to clear buffers, so there is no need to enable libcryptoc.

BUG=b:277029648
TEST=make BOARD=hammer
TEST=make BOARD=prism
TEST=make BOARD=bloonchipper
TEST=make BOARD=dartmonkey

Change-Id: Ibc1e276e221a015285f3b1946df23510a578577d