charger: Rename enum charge_state to led_pwr_state
This name is confusing, since its members start with PWR_ and the
comment for enum charge_state_v2 says that these values are only sued
to control the LEDs, with one not-quite-correct exception, whatever
that might be.

We could rename it to power_state, but that is already fairly widely
used as a variable in the code base. Also it seems better to match the
enum PWR_ prefix. Future work may rename the members too.

Rename the charge_state enum.

This makes no functional change.

BUG=b:218332694
TEST=make buildall

Change-Id: If8a7cc73dcfb4c1e89984e57fc55e432416e8aec